+++ server/config/cache.php
@@ -1,0 +1,18 @@
+<?php
+
+return [
+    'default' => env('CACHE_MODE', 'file'),
+    'stores' => [
+        'file' => [
+            'driver' => 'file',
+            'path' => runtime_path('cache')
+        ],
+        'redis' => [
+            'driver' => 'redis',
+            'connection' => 'default'
+        ],
+        'array' => [
+            'driver' => 'array'
+        ]
+    ]
+];